+++ src/main/java/io/kuzzle/sdk/core/KuzzleDataCollection.java
@@ -36,10 +36,10 @@ public class KuzzleDataCollection {
    * or like a room for pub/sub messages.
    *
    * @param kuzzle     the kuzzle
-   * @param index      the index
    * @param collection the collection
+   * @param index      the index
    */
-  public KuzzleDataCollection(@NonNull final Kuzzle kuzzle, @NonNull final String index, @NonNull final String collection) {
+  public KuzzleDataCollection(@NonNull final Kuzzle kuzzle, @NonNull final String collection, @NonNull final String index) {
     if (kuzzle == null) {
       throw new IllegalArgumentException("KuzzleDataCollection: need a Kuzzle instance to initialize");
     }
@@ -67,10 +67,9 @@ public class KuzzleDataCollection {
    *
    * @param filter   the filter
    * @param listener the listener
-   * @return the kuzzle data collection
    */
-  public KuzzleDataCollection advancedSearch(final JSONObject filter, final KuzzleResponseListener<KuzzleDocumentList> listener) {
-    return this.advancedSearch(filter, null, listener);
+  public void advancedSearch(final JSONObject filter, final KuzzleResponseListener<KuzzleDocumentList> listener) {
+    this.advancedSearch(filter, null, listener);
   }
 
   /**
@@ -82,9 +81,8 @@ public class KuzzleDataCollection {
    * @param filters  the filters
    * @param options  the options
    * @param listener the listener
-   * @return KuzzleDataCollection object
    */
-  public KuzzleDataCollection advancedSearch(final JSONObject filters, final KuzzleOptions options, @NonNull final KuzzleResponseListener<KuzzleDocumentList> listener) {
+  public void advancedSearch(final JSONObject filters, final KuzzleOptions options, @NonNull final KuzzleResponseListener<KuzzleDocumentList> listener) {
     if (listener == null) {
       throw new IllegalArgumentException("listener cannot be null");
     }
@@ -123,7 +121,6 @@ public class KuzzleDataCollection {
     } catch (JSONException e) {
       throw new RuntimeException(e);
     }
-    return this;
   }
 
   /**
@@ -147,20 +144,18 @@ public class KuzzleDataCollection {
    *
    * @param filters  the filters
    * @param listener the listener
-   * @return the kuzzle data collection
    */
-  public KuzzleDataCollection count(final JSONObject filters, @NonNull final KuzzleResponseListener<Integer> listener) {
-    return this.count(filters, null, listener);
+  public void count(final JSONObject filters, @NonNull final KuzzleResponseListener<Integer> listener) {
+    this.count(filters, null, listener);
   }
 
   /**
    * Count kuzzle data collection.
    *
    * @param listener the listener
-   * @return the kuzzle data collection
    */
-  public KuzzleDataCollection count(@NonNull final KuzzleResponseListener<Integer> listener) {
-    return this.count(null, null, listener);
+  public void count(@NonNull final KuzzleResponseListener<Integer> listener) {
+    this.count(null, null, listener);
   }
 
   /**
@@ -172,9 +167,8 @@ public class KuzzleDataCollection {
    * @param filters  the filters
    * @param options  the options
    * @param listener the cb
-   * @return KuzzleDataCollection kuzzle data collection
    */
-  public KuzzleDataCollection count(final JSONObject filters, final KuzzleOptions options, @NonNull final KuzzleResponseListener<Integer> listener) {
+  public void count(final JSONObject filters, final KuzzleOptions options, @NonNull final KuzzleResponseListener<Integer> listener) {
     if (listener == null) {
       throw new IllegalArgumentException("KuzzleDataCollection.count: listener required");
     }
@@ -200,7 +194,6 @@ public class KuzzleDataCollection {
     } catch (JSONException e) {
       throw new RuntimeException(e);
     }
-    return this;
   }
 
   /**
@@ -687,10 +680,9 @@ public class KuzzleDataCollection {
    *
    * @param documentId the document id
    * @param listener   the listener
-   * @return the kuzzle data collection
    */
-  public KuzzleDataCollection fetchDocument(@NonNull final String documentId, @NonNull final KuzzleResponseListener<KuzzleDocument> listener) {
-    return this.fetchDocument(documentId, null, listener);
+  public void fetchDocument(@NonNull final String documentId, @NonNull final KuzzleResponseListener<KuzzleDocument> listener) {
+    this.fetchDocument(documentId, null, listener);
   }
 
   /**
@@ -699,9 +691,8 @@ public class KuzzleDataCollection {
    * @param documentId the document id
    * @param options    the options
    * @param listener   the listener
-   * @return KuzzleDataCollection kuzzle data collection
    */
-  public KuzzleDataCollection fetchDocument(@NonNull final String documentId, final KuzzleOptions options, final KuzzleResponseListener<KuzzleDocument> listener) {
+  public void fetchDocument(@NonNull final String documentId, final KuzzleOptions options, final KuzzleResponseListener<KuzzleDocument> listener) {
     if (documentId == null) {
       throw new IllegalArgumentException("KuzzleDataCollection.fetchDocument: documentId required");
     }
@@ -735,17 +726,15 @@ public class KuzzleDataCollection {
     } catch (JSONException e) {
       throw new RuntimeException(e);
     }
-    return this;
   }
 
   /**
    * Retrieves all documents stored in this data collection.
    *
    * @param listener the listener
-   * @return the kuzzle data collection
    */
-  public KuzzleDataCollection fetchAllDocuments(@NonNull final KuzzleResponseListener<KuzzleDocumentList> listener) {
-    return this.fetchAllDocuments(null, listener);
+  public void fetchAllDocuments(@NonNull final KuzzleResponseListener<KuzzleDocumentList> listener) {
+    this.fetchAllDocuments(null, listener);
   }
 
   /**
@@ -753,9 +742,8 @@ public class KuzzleDataCollection {
    *
    * @param options  the options
    * @param listener the listener
-   * @return the kuzzle data collection
    */
-  public KuzzleDataCollection fetchAllDocuments(final KuzzleOptions options, @NonNull final KuzzleResponseListener<KuzzleDocumentList> listener) {
+  public void fetchAllDocuments(final KuzzleOptions options, @NonNull final KuzzleResponseListener<KuzzleDocumentList> listener) {
     if (listener == null) {
       throw new IllegalArgumentException("KuzzleDataCollection.fetchAllDocuments: listener required");
     }
@@ -772,17 +760,16 @@ public class KuzzleDataCollection {
         throw new RuntimeException(e);
       }
     }
-    return this.advancedSearch(filter, options, listener);
+    this.advancedSearch(filter, options, listener);
   }
 
   /**
    * Instantiates a KuzzleDataMapping object containing the current mapping of this collection.
    *
    * @param listener the listener
-   * @return the mapping
    */
-  public KuzzleDataCollection getMapping(@NonNull final KuzzleResponseListener<KuzzleDataMapping> listener) {
-    return this.getMapping(null, listener);
+  public void getMapping(@NonNull final KuzzleResponseListener<KuzzleDataMapping> listener) {
+    this.getMapping(null, listener);
   }
 
   /**
@@ -790,14 +777,12 @@ public class KuzzleDataCollection {
    *
    * @param options  the options
    * @param listener the listener
-   * @return the mapping
    */
-  public KuzzleDataCollection getMapping(final KuzzleOptions options, @NonNull final KuzzleResponseListener<KuzzleDataMapping> listener) {
+  public void getMapping(final KuzzleOptions options, @NonNull final KuzzleResponseListener<KuzzleDataMapping> listener) {
     if (listener == null) {
       throw new IllegalArgumentException("KuzzleDataCollection.getMapping: listener required");
     }
     new KuzzleDataMapping(this).refresh(options, listener);
-    return this;
   }
 
   /**