+++ CHANGELOG.md
@@ -1,12 +1,23 @@
 *__note:__ the # at the end of lines are the pull request numbers on GitHub*
 
+# Current
+
+## Breaking Changes
+
+* `KuzzleDataCollection` constructor signature has been changed from:  
+`KuzzleDataCollection(kuzzle, index, collection)`  
+ to:  
+`KuzzleDataCollection(kuzzle, collection, index)`  
+This has been done to make it on par with the `Kuzzle.dataCollectionFactory` method
+
+
 # 1.8.0
 
 * https://github.com/kuzzleio/sdk-android/releases/tag/1.8.0
 
 # 1.7.0
 
-* https://github.com/kuzzleio/sdk-android/releases/tag/1.7.0 
+* https://github.com/kuzzleio/sdk-android/releases/tag/1.7.0
 
 # 1.6.1